+++ mainjs/utiles.js
@@ -30,10 +30,12 @@ if (!localStorage.getItem('listaUtiles')) {
             localStorage.setItem('listaUtiles', JSON.stringify(data))
             listaUtiles = JSON.parse(localStorage.getItem('listaUtiles'))
             console.log(listaUtiles)
+            mostrarSocios()
         })
 } else if (localStorage.getItem('listaUtiles')) {
     listaUtiles = JSON.parse(localStorage.getItem('listaUtiles'))
     console.log(listaUtiles)
+    mostrarSocios()
 }
 
 myFormUtiles.addEventListener('submit', (event) => {
@@ -77,17 +79,65 @@ myFormUtiles.addEventListener('submit', (event) => {
         listaUtiles.push(socio)
     }
 
-    if (inputJardin.checked) {
-        crearSocio(inputNumFuncUtiles.value, inputJardin.value, inputEdad.value)
-    } else if (inputPrimaria.checked) {
-        crearSocio(inputNumFuncUtiles.value, inputPrimaria.value, inputEdad.value)
-    } else if (inputSecCicloBas.checked) {
-        crearSocio(inputNumFuncUtiles.value, inputSecCicloBas.value, inputEdad.value)
-    } else if (inputSecBachi.checked) {
-        crearSocio(inputNumFuncUtiles.value, inputSecBachi.value, inputEdad.value)
-    }
-    localStorage.setItem('listaUtiles', JSON.stringify(listaUtiles))
-    myFormUtiles.reset()
-    console.log(listaUtiles)
+    const swalWithBootstrapButtons = Swal.mixin({
+        customClass: {
+            confirmButton: 'btn btn-success',
+            cancelButton: 'btn btn-danger'
+        },
+        buttonsStyling: false
+    })
+    
+    swalWithBootstrapButtons.fire({
+        title: 'Confirma la solicitud?',
+        icon: 'question',
+        showCancelButton: true,
+        confirmButtonText: 'Aceptar',
+        cancelButtonText: 'Cancelar',
+        reverseButtons: true
+    }).then((result) => {
+        if (result.isConfirmed) {
+            swalWithBootstrapButtons.fire(
+                'Solicitud realizada.',
+                'Para confirmar que su solicitud ha sido ingresada correctamente recargue la página, debería figurar en la lista de solicitudes confirmadas.'
+            )
+            if (inputJardin.checked) {
+                crearSocio(inputNumFuncUtiles.value, inputJardin.value, inputEdad.value)
+            } else if (inputPrimaria.checked) {
+                crearSocio(inputNumFuncUtiles.value, inputPrimaria.value, inputEdad.value)
+            } else if (inputSecCicloBas.checked) {
+                crearSocio(inputNumFuncUtiles.value, inputSecCicloBas.value, inputEdad.value)
+            } else if (inputSecBachi.checked) {
+                crearSocio(inputNumFuncUtiles.value, inputSecBachi.value, inputEdad.value)
+            }
+            localStorage.setItem('listaUtiles', JSON.stringify(listaUtiles))
+            myFormUtiles.reset()
+            console.log(listaUtiles)
+    
+        } else if (
+            /* Read more about handling dismissals below */
+            result.dismiss === Swal.DismissReason.cancel
+        ) {
+            swalWithBootstrapButtons.fire(
+                'Cancelar',
+                'La solicitud ha sido cancelada.'
+            )
+        }
+    })
+    
+
 })
-// Agregar un fetch que permita corroborar si el numero de socio se encuentra en la lista de socios activos.
+
+
+function mostrarSocios() {
+    let listaUtilesParse = JSON.parse(localStorage.getItem('listaUtiles'))
+    console.log(listaUtilesParse);
+
+    let containerUtiles = document.querySelector('#Solicitudes')
+    listaUtilesParse.forEach(socio => {
+        containerUtiles.innerHTML += `<p>- el número de socio: ${socio.numero}, ha solicitado útiles de ${socio.nivel_de_estudio}.</p>`
+    })
+    
+    }
+    
+
+
